requirements.txt
# TradingAgents OpenAI Framework Dependencies

# 核心依赖
openai>=1.3.0
python-dotenv>=0.19.0

# 数据处理
pandas>=1.5.0
numpy>=1.21.0
yfinance>=0.2.65

# 网络请求和爬虫
requests>=2.28.0
httpx>=0.24.0
beautifulsoup4>=4.11.0
selenium>=4.8.0

# 重试机制
tenacity>=8.2.0

# 数据存储
json5>=0.9.0

# 日志和监控
loguru>=0.6.0

# 测试
pytest>=7.0.0
pytest-asyncio>=0.21.0

# 工具
python-dateutil>=2.8.0
tqdm>=4.64.0

# 数据源API
newsapi-python>=0.2.6
praw>=7.6.0  # Reddit API
tweepy>=4.14.0  # Twitter API

# 技术分析
TA-Lib>=0.4.29  # 需要系统级C库支持，可选安装
mplfinance>=0.12.9b7  # 金融图表，使用可用的beta版本
stockstats>=0.5.0  # 技术指标计算

# Excel支持
openpyxl>=3.0.0
xlrd>=2.0.0